+++ credit-tree/package.json
@@ -7,8 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "jest",
-    "test:watch": "jest --watch"
+    "test": "vitest"
   },
   "dependencies": {
     "@prisma/client": "6.10.1",
@@ -21,6 +20,7 @@
     "@types/three": "0.177.0",
     "@types/validator": "13.15.2",
     "bcrypt": "6.0.0",
+    "dotenv": "17.2.0",
     "next": "15.3.2",
     "postprocessing": "6.37.4",
     "r3f-perf": "7.2.3",
@@ -31,20 +31,23 @@
     "validator": "13.15.15"
   },
   "devDependencies": {
+    "@babel/preset-env": "7.28.0",
+    "@babel/preset-react": "7.27.1",
+    "@babel/preset-typescript": "7.27.1",
     "@eslint/eslintrc": "^3",
     "@testing-library/dom": "10.4.0",
-    "@testing-library/jest-dom": "6.6.3",
     "@testing-library/react": "16.3.0",
-    "@types/jest": "30.0.0",
     "@types/node": "20.19.1",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "@vitejs/plugin-react": "4.6.0",
     "eslint": "^9",
     "eslint-config-next": "15.3.2",
-    "jest": "30.0.3",
-    "jest-environment-jsdom": "30.0.2",
+    "jsdom": "26.1.0",
     "prisma": "6.10.1",
     "ts-node": "10.9.2",
-    "typescript": "5.8.3"
+    "typescript": "5.8.3",
+    "vite-tsconfig-paths": "5.1.4",
+    "vitest": "3.2.4"
   }
 }
