+++ credit-tree/vitest.config.mts
@@ -1,0 +1,15 @@
+import { defineConfig } from "vitest/config";
+import react from "@vitejs/plugin-react";
+import tsconfigPaths from "vite-tsconfig-paths";
+import { config } from "dotenv";
+
+export default defineConfig({
+  plugins: [tsconfigPaths(), react()],
+  test: {
+    environment: "jsdom",
+    globals: true,
+    env: {
+      ...config({ path: "./.env" }).parsed,
+    },
+  },
+});
